fix: support half-step importance/urgency scores (0-4, step 0.5) (GREEN)
- models.rs: Plan/CreatePlanRequest/UpdatePlanRequest 的 importance/urgency
  i32 → f64，serde 接受 2.5 等半格值
- db.rs: migration 3 重建 plans 表，列类型 INTEGER → REAL（SQLite 无法
  ALTER COLUMN TYPE；事务内关外键重建，plan_logs 的 FK 字符串引用自动
  指向新表，旧数据完整迁移）
- plans.rs: create_plan/update_plan 增加 0-4 范围校验（review 🟡 建议）

验证：cargo test 22 passed, 0 failed（含 4 个新测试）

AI[100%] Human Fixed[0%] Human[0%] AI Adopted[100%]

+++ app/src-tauri/src/models.rs
@@ -31,8 +31,8 @@ pub struct Plan {
     pub description: String,
     pub category_id: Option<String>,
     pub parent_id: Option<String>,
-    pub importance: i32,
-    pub urgency: i32,
+    pub importance: f64,
+    pub urgency: f64,
     pub ddl: Option<String>,
     pub tag_workflow_id: Option<String>,
     pub current_step_index: i32,
@@ -64,9 +64,9 @@ pub struct CreatePlanRequest {
     pub category_id: Option<String>,
     pub parent_id: Option<String>,
     #[serde(default)]
-    pub importance: i32,
+    pub importance: f64,
     #[serde(default)]
-    pub urgency: i32,
+    pub urgency: f64,
     pub ddl: Option<String>,
     pub tag_workflow_id: Option<String>,
     #[serde(default)]
@@ -93,8 +93,8 @@ pub struct UpdatePlanRequest {
     pub description: Option<String>,
     pub category_id: Option<Option<String>>,
     pub parent_id: Option<Option<String>>,
-    pub importance: Option<i32>,
-    pub urgency: Option<i32>,
+    pub importance: Option<f64>,
+    pub urgency: Option<f64>,
     pub ddl: Option<Option<String>>,
     pub tag_workflow_id: Option<Option<String>>,
     pub current_step_index: Option<i32>,
